+++ db/migrate/20150528113733_change_integer_columns_purpose_codes.rb
@@ -1,7 +1,7 @@
 class ChangeIntegerColumnsPurposeCodes < ActiveRecord::Migration
   def change
-    change_column :purpose_codes, :txn_limit, :float, :precision => 0
-    change_column :purpose_codes, :mtd_txn_cnt_self, :float, :precision => 0
-    change_column :purpose_codes, :mtd_txn_cnt_sp, :float, :precision => 0
+    change_column :purpose_codes, :txn_limit, :decimal, :precision => 0
+    change_column :purpose_codes, :mtd_txn_cnt_self, :decimal, :precision => 0
+    change_column :purpose_codes, :mtd_txn_cnt_sp, :decimal, :precision => 0
   end
 end
