+++ db/migrate/20150602085257_add_bal_available_to_inward_remittances.rb
@@ -1,5 +1,5 @@
 class AddBalAvailableToInwardRemittances < ActiveRecord::Migration
   def change
-    add_column :inward_remittances, :bal_available, :float
+    add_column :inward_remittances, :bal_available, :decimal
   end
 end
